pyproject.toml
[project]
name = "my-cner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "google-genai>=1.56.0",
    "lxml>=6.0.2",
    "pyidaungsu>=0.1.4",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "selenium>=4.39.0",
    "textual>=6.11.0",
    "tqdm>=4.67.1",
    "webdriver-manager>=4.0.2",
]
